Make AppABI an inner class
