+++ app-android/app/src/main/kotlin/com/algoritmico/passepartout/helpers/Helpers.kt
@@ -9,3 +9,19 @@ import kotlinx.serialization.json.Json
 val globalJsonCoder = Json {
     ignoreUnknownKeys = true
 }
+
+class ABIException(
+    val code: Int,
+    val payload: String?
+) : RuntimeException("ABI call failed (code=$code): $payload")
+
+data class ABIResult(
+    val code: Int,
+    val payload: String?
+) {
+    fun getOrThrow() {
+        if (code != 0) {
+            throw ABIException(code, payload)
+        }
+    }
+}
